pyproject.toml
[project]
name = "lbm-training-framework"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.6",
    "django-ninja>=1.4.3",
]
